+++ flow_py_library_general_maths/general_maths/add/flow_add.py
@@ -1,5 +1,5 @@
 from flow import Component, LogLevel
-from flow_types import base
+from flow_types import base, unions
 
 inports = ["val1", "val2"]
 outports = ["result"]
@@ -8,17 +8,17 @@ definition = {
     "name": "add",
     "description": "Adds two numbers together.",
     "inports": [
-        {"name": inports[0], "description": "The first number", "types": [base.Double]},
-        {"name": inports[1], "description": "The second number", "types": [base.Double]},
+        {"name": inports[0], "description": "The first number", "types": unions.Number},
+        {"name": inports[1], "description": "The second number", "types": unions.Number},
     ],
-    "outports": [{"name": outports[0], "description": "The result number", "types": [base.Double]}],
+    "outports": [{"name": outports[0], "description": "The result number", "types": unions.Number}],
 }
 
 
 # The process that the component performs
 def process(component: Component):
     # check that the components have data --> this can be modified if you want to set explicit defaults etc.
-    if not (component.has_data(inports[0]) and component.has_data(inports[1])):
+    if not component.has_data():
         return
 
     # source the data from the inports
@@ -36,5 +36,5 @@ def process(component: Component):
     if component.debug:
         component.log(log_level=LogLevel.DEBUG, message=f"The result is {result}.")
 
-    # send the result message to the outports (as addressable)
-    component.send_data_addressable(result_msg, outports[0])
+    # send the result message to the outports (as multi_connection)
+    component.send_data(result_msg, outports[0])