+++ flow_py_library_general_maths/general_maths/e_ln/flow_e_ln.py
@@ -1,7 +1,7 @@
 import math
 
 from flow import Component, LogLevel
-from flow_types import base
+from flow_types import base, unions
 
 inports = ["value"]
 outports = ["result"]
@@ -9,8 +9,8 @@ outports = ["result"]
 definition = {
     "name": "natural_log",
     "description": "Natural log of the value.",
-    "inports": [{"name": inports[0], "description": "The natural log of the value", "types": [base.Double]}],
-    "outports": [{"name": outports[0], "description": "The result number", "types": [base.Double]}],
+    "inports": [{"name": inports[0], "description": "The natural log of the value", "types": unions.Number}],
+    "outports": [{"name": outports[0], "description": "The result number", "types": unions.Number}],
 }
 
 
@@ -33,5 +33,5 @@ def process(component: Component):
     if component.debug:
         component.log(log_level=LogLevel.DEBUG, message=f"The result is {result}.")
 
-    # send the result message to the outports (as addressable)
-    component.send_data_addressable(result_msg, outports[0])
+    # send the result message to the outports (as multi_connection)
+    component.send_data(result_msg, outports[0])