+++ flow_py_library_general_maths/general_maths/e_log_n/flow_e_log_n.py
@@ -1,7 +1,7 @@
 import math
 
 from flow import Component, LogLevel
-from flow_types import base
+from flow_types import base, unions
 
 inports = ["index", "base"]
 outports = ["result"]
@@ -12,17 +12,17 @@ definition = {
     "description": "Raises an index to a power and returns the result."
     + "(Negative index values are treated as positive with negatived result).",
     "inports": [
-        {"name": inports[0], "description": "The index to log, log(I, b)", "types": [base.Double]},
-        {"name": inports[1], "description": "The base to log, log(i, B)", "types": [base.Double]},
+        {"name": inports[0], "description": "The index to log, log(I, b)", "types": unions.Number},
+        {"name": inports[1], "description": "The base to log, log(i, B)", "types": unions.Number},
     ],
-    "outports": [{"name": outports[0], "description": "The result number", "types": [base.Double]}],
+    "outports": [{"name": outports[0], "description": "The result number", "types": unions.Number}],
 }
 
 
 # The process that the component performs
 def process(component: Component):
     # check that the components have data --> this can be modified if you want to set explicit defaults etc.
-    if not (component.has_data(inports[0]) and component.has_data(inports[1])):
+    if not component.has_data():
         return
 
     # source the data from the inports
@@ -40,5 +40,5 @@ def process(component: Component):
     if component.debug:
         component.log(log_level=LogLevel.DEBUG, message=f"The result is {result}.")
 
-    # send the result message to the outports (as addressable)
-    component.send_data_addressable(result_msg, outports[0])
+    # send the result message to the outports (as multi_connection)
+    component.send_data(result_msg, outports[0])