+++ tests/arrays/test_average.py
@@ -1,37 +1,38 @@
 import numpy as np
 from flow.test_framework import FlowTest, flow_test
+from flow.test_framework.helpers import assert_test_data_expected
 from flow_types import base
 
-from flow_py_library_general_maths.array.sum.flow_sum import inports, outports
-from flow_py_library_general_maths.util.utils_tests import basic_test_eval
+from flow_py_library_general_maths.array.average.flow_average import inports, outports
 
 # Tests
-component_file = "array_maths/sum"
+component_file = "array_maths/average"
 
 
 def run_test_func(inputs, outputs, flow: FlowTest):
     test_data = flow.test(component_file, inputs, outputs)
-    basic_test_eval(test_data)
+    assert_test_data_expected(test_data)
 
 
 def test_MdDouble2x2(flow: FlowTest):
     array = base.MdDouble(np.array([[1, 2], [3, 4]]))
-
     inputs = {inports[0]: [array]}
 
-    outputs = {outports[0]: base.Double(10)}
+    outputs = {outports[0]: base.Double(2.5)}
 
     run_test_func(inputs, outputs, flow=flow)
 
 
 def test_MdDouble5x5x5x5x5(flow: FlowTest):
     array = base.MdDouble(
-        np.array([[1, 2, 3, 4, 5], [6, 7, 8, 9, 10], [11, 12, 13, 14, 15], [16, 17, 18, 19, 20], [21, 22, 23, 24, 25]]),
+        np.array(
+            [[1, 2, 3, 4, 5], [6, 7, 8, 9, 10], [11, 12, 13, 14, 15], [16, 17, 18, 19, 20], [21, 22, 23, 24, 25]],
+        ),
     )
 
     inputs = {inports[0]: [array]}
 
-    outputs = {outports[0]: base.Double(325)}
+    outputs = {outports[0]: base.Double(13)}
 
     run_test_func(inputs, outputs, flow=flow)