+++ tests/arrays/test_linspace.py
@@ -2,6 +2,7 @@ from typing import List
 
 import pytest
 from flow.test_framework import FlowTest, flow_test
+from flow.test_framework.helpers import assert_test_data_expected
 from flow_types import base
 
 from flow_py_library_general_maths.array.assemble.linspace.flow_linspace import (
@@ -10,15 +11,14 @@ from flow_py_library_general_maths.array.assemble.linspace.flow_linspace import 
     start_port,
     stop_port,
 )
-from flow_py_library_general_maths.util.utils_tests import basic_test_eval
 
 # Tests
 component_file = "array_maths/assemble/linspace"
 
 
 def run_test_func(inputs, outputs, flow: FlowTest):
-    test_data = flow.test(component_file, inputs, outputs, timeout=1000)
-    basic_test_eval(test_data)
+    test_data = flow.test(component_file, inputs, outputs)
+    assert_test_data_expected(test_data)
 
 
 @pytest.mark.parametrize(