+++ tests/general_maths/test_e_ln.py
@@ -1,18 +1,18 @@
 import math
 
 from flow.test_framework import FlowTest, flow_test
+from flow.test_framework.helpers import assert_test_data_expected
 from flow_types import base
 
 from flow_py_library_general_maths.general_maths.e_ln.flow_e_ln import inports, outports
-from flow_py_library_general_maths.util.utils_tests import basic_test_eval
 
 # Tests
-component_file = "general_maths/e_ln"
+component_file = "maths/e_ln"
 
 
 def run_test_func(inputs, outputs, flow: FlowTest):
     test_data = flow.test(component_file, inputs, outputs)
-    basic_test_eval(test_data)
+    assert_test_data_expected(test_data)
 
 
 def test_ln_double(flow: FlowTest):