+++ tests/general_maths/test_subtract.py
@@ -1,29 +1,29 @@
 from flow.test_framework import FlowTest, flow_test
+from flow.test_framework.helpers import assert_test_data_expected
 from flow_types import base
 
-from flow_py_library_general_maths.general_maths.divide.flow_divide import inports, outports
-from flow_py_library_general_maths.util.utils_tests import basic_test_eval
+from flow_py_library_general_maths.general_maths.subtract.flow_subtract import inports, outports
 
 # Tests
-component_file = "general_maths/divide"
+component_file = "maths/subtract"
 
 
 def run_test_func(inputs, outputs, flow: FlowTest):
     test_data = flow.test(component_file, inputs, outputs)
-    basic_test_eval(test_data)
+    assert_test_data_expected(test_data)
 
 
-def test_divide_double2double(flow: FlowTest):
+def test_subtract_double2double(flow: FlowTest):
     inputs = {
         inports[0]: [base.Double(3.67)],
         inports[1]: [base.Double(-1e9)],
     }
 
-    outputs = {outports[0]: [base.Double(3.67 / (-1e9))]}
+    outputs = {outports[0]: [base.Double(3.67 + 1e9)]}
 
     run_test_func(inputs, outputs, flow=flow)
 
 
 if __name__ == "__main__":
     with flow_test() as flow:
-        test_divide_double2double(flow)
+        test_subtract_double2double(flow)